Added rate-coded input-output analysis for the golgi network.
Added groups for golgi cells and for a new class of spike relays (which are necessary to have both AMPA and NMDA components in external simulation of GoCs), network connections (GJs and dummy relays->GoCs) and scripts to automate the simulation for the different values of stimulation rate. Also, added a plotting script for the I/O relationship.